v0.11.0-A01
Mountain Guarder v0.11.0 Alpha 01

I know it won't update because of the heroku incident but it will be released anyways for the sake of version control.

Features Adicytes store, inventory changes, and performance optimizations.
The FPS boost increases average FPS, but it makes the game stutter really badly and sometimes it just hitches so bad it feels like a lower 20fps with more consistency is better.